+++ OpenHD/ohd_video/CMakeLists.txt
@@ -50,7 +50,7 @@ find_package(Threads REQUIRED)
 
 set(sources 
     "inc/camerastream.h"
-    "inc/DCameras.h"
+    "inc/camera_discovery.h"
     "inc/gstreamerstream.h"
     "inc/libcamera_provider.hpp"
     "inc/OHDGstHelper.hpp"
@@ -63,11 +63,11 @@ set(sources
     "inc/v_validate_settings.h"
     "inc/veyestream.h"
 
-    "src/DCamerasHelper.hpp"
+    "src/camera_discovery_helper.hpp"
     "src/ffmpeg_videosamples.hpp"
 
     "src/camerastream.cpp"
-    "src/DCameras.cpp"
+    "src/camera_discovery.cpp"
     "src/gstreamerstream.cpp"
     "src/OHDVideo.cpp"
     "src/veyestream.cpp"
