+++ OpenHD/ohd_video/src/camera_discovery_helper.hpp
