+++ OpenHD/ohd_video/test/test_video.cpp
@@ -1,14 +1,14 @@
 
+#include <camera_discovery.h>
+
+#include <chrono>
 #include <iostream>
 #include <thread>
-#include <chrono>
-#include <DCameras.h>
-
-#include "openhd-profile.hpp"
-#include "openhd-platform.hpp"
-#include "openhd-platform-discover.hpp"
 
 #include "OHDVideo.h"
+#include "openhd-platform-discover.hpp"
+#include "openhd-platform.hpp"
+#include "openhd-profile.hpp"
 
 int main(int argc, char *argv[]) {
   const auto platform=DPlatform::discover();
